Move new data classes to relevant packages
