+++ collect_app/src/main/java/org/odk/collect/android/storage/StoragePaths.kt
@@ -1,0 +1,8 @@
+package org.odk.collect.android.storage
+
+data class StoragePaths(
+    val rootDir: String,
+    val formsDir: String,
+    val instancesDir: String,
+    val cacheDir: String
+)
